pyproject.toml
[project]
name = "ollama-obsidian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "faiss-cpu>=1.11.0.post1",
    "fast-agent-mcp>=0.2.46",
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "markdown2>=2.5.3",
    "mcp[cli]>=1.12.1",
    "ollmcp>=0.15.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "unstructured>=0.18.11",
    "uvicorn>=0.35.0",
]
